# specs/001-doc-assistant-rag/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Doc Assistant API
  version: "0.1.0"
  description: |
    RAG-powered document Q&A. All endpoints (except `/healthz` and `/metrics`)
    require a bearer token (`APP_SHARED_TOKEN` on the server).
servers:
  - url: http://localhost:8000
    description: Local dev
security:
  - BearerToken: []
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: Single shared secret loaded from `APP_SHARED_TOKEN`.
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - unauthorized
            - bad_request
            - unsupported_media_type
            - payload_too_large
            - not_found
            - upstream_error
            - upstream_unavailable
            - rate_limited        # placeholder; v1 does not enforce
            - internal_error
        message:
          type: string
          description: Human-readable, safe to render in UI. Never contains secrets or stack traces.
        request_id:
          type: string
...
    Citation:
      type: object
      required: [chunk_id, document_id, locator, score]
      properties:
        chunk_id:
          type: string
        document_id:
          type: string
          format: uuid
        locator:
          type: string
          description: 'Human-readable: "page 12" (PDF) or "Article 5 > §2" (DOCX).'
        score:
          type: number
          format: float
    HistoryResponse:
      type: object
      required: [session_id, turns]
      properties:
        session_id:
          type: string
        turns:
          type: array
          items:
            $ref: "#/components/schemas/ConversationTurn"
    ConversationTurn:
      type: object
      required: [turn_id, role, content, created_at]
      properties:
        turn_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        citations:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Citation"
        created_at:
          type: string
          format: date-time
        state:
          type: string
          nullable: true
          enum: [complete, stopped, errored]
...
paths:
...
  /history/{session_id}:
    get:
      summary: Fetch the ordered conversation history for a session.
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: History.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Session not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...